requirements.txt
torch>=1.9.0
torchvision>=0.10.0
faiss-cpu>=1.7.0  # or faiss-gpu for GPU support
numpy>=1.19.0
Pillow>=8.0.0
matplotlib>=3.3.0
open-clip-torch>=2.0.0
tqdm>=4.50.0
scikit-learn>=0.24.0
# Install medmnist package
pip install medmnist

# Install other required packages if not already installed
pip install torch torchvision transformers faiss-cpu matplotlib numpy pillow tqdm scikit-learn

# In a notebook cell
!pip install medmnist
!pip install torch torchvision transformers faiss-cpu matplotlib numpy pillow tqdm scikit-learn
